+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Develop
 
+- Introduced `import_fields`: a subroutine to read and import fld data,
+  with interpolation capabilities. 
 - Added `vector_list_t` and `name` to `vector_t`.
 - Rework hash table iterators, significantly faster (O(tsize) => O(entries)
 - Remove redundant directory in `site-packages` when installing pyneko